chore: remove scenario_engine
