+++ src/rai/rai/tools/ros/native.py
@@ -30,7 +30,10 @@ from langchain.tools import BaseTool
 from langchain_core.pydantic_v1 import BaseModel, Field
 from rclpy.impl.rcutils_logger import RcutilsLogger
 
-from rai.node import RaiNode
+from typing import TYPE_CHECKING
+
+if TYPE_CHECKING:
+    from rai.node import RaiNode
 
 from .utils import convert_ros_img_to_base64, import_message_from_str
 
@@ -71,7 +74,7 @@ class PubRos2MessageToolInput(BaseModel):
 # --------------------- Tools ---------------------
 class Ros2BaseTool(BaseTool):
     # TODO: Make the decision between rclpy.node.Node and RaiNode
-    node: rclpy.node.Node | RaiNode = Field(
+    node: "rclpy.node.Node | RaiNode" = Field(
         ..., exclude=True, include=False, required=True
     )
